src/main/java/com/lqy/sell/service/impl/ProductCategoryServiceImpl.java
package com.lqy.sell.service.impl;

import com.lqy.sell.dataobject.ProductCategory;
import com.lqy.sell.repository.ProductCategoryRepository;
import com.lqy.sell.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * 类目Service实现
 * Created by Rodriguez
 * 2018/7/12 9:40
 */
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer productCategoryId) {
        return repository.findOne(productCategoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }

}
